+++ src/core/syswarden-cli/cmd/install.go
@@ -117,7 +117,7 @@ var installCmd = &cobra.Command{
 		}
 		ip := getPublicIP()
 		fmt.Printf("\n======================================================\n")
-		fmt.Printf("[+] URL d'accès client Web-TUI : http://%s:%s/?token=%s\n", ip, webtuiPort, token)
+		fmt.Printf("[+] URL d'accès client Web-TUI : https://%s:%s/?token=%s\n", ip, webtuiPort, token)
 		fmt.Printf("======================================================\n\n")
 
 		fmt.Println("[SYSWARDEN] v3.72.1 Native Installation Complete.")